Fixed ThreadInfo behavior
The previous implementation was setting the values of s_threads based on an index, and for unknown reasons this caused an IndexOutOfRangeException to be thrown. Note that this exception was not raised from within the program, but rather external .dll files.
Now, whenever a new ThreadInfo object is created, the constructor stores the object inside the s_threads array in the first available spot.
This fixed all issues with the previous implementation, but it might need to be revised in the future as the ThreadInfo constructor may now have too many responsibilities, which could be better addressed somewhere else.