Implemented repetition table
Added a repetition table in place of the PositionKeyHistory stack with the same purpose.
Keys are stored in an array with at most 256 entries.
Checking for a draw by repetition is more efficient because we only look at entries that have the same side to move and only up to the last irreversible move made.
Overall, this is still slower than the previous implementation.