Added upcoming repetition detection
Added HasUpcomingRepetition to the RepetionTable to check if a move in the current position would lead to a repetition. This function does not generate the legal moves in this position to check for this. Instead, it uses an algorithm found in Stockfish from the following article by Marcel van Kervinck: http://web.archive.org/web/20201107002606/https://marcelk.net/2013-04-06/paper/upcoming-rep-v2.pdf.
Currently, the search is significantly slower than before, but this is likely because of a bug. Once everything works properly it is expected to result in a very small slow down, with an improved performance.
The reason this function was implemented is to avoid returning TT cutoffs in a position where the TT move would now result in a repetition of moves instead of a positive score.