Added PlyCountReversible
Added a separate ply counter for reversible moves, which unlike FiftyMovePlyCount is reset when making a null move.
Theoretically, using this counter when checking for draws by repetition should be faster, but currently that's not the case.
Only if null moves do not reset the ply counter used, the performance results are acceptable.
It is unclear why this is the case since any positions before a null move was made should be unreachable and thus just slow the algorithm down.