Correct implementation of useNullMovePruning
The useNullMovePruning parameter is supposed to disable null moves inside a branch where one was already previously performed.
However, the implementation for this, passing the value down the search tree, was missing.
Adding this makes the program more stable with regards to the new draw by repetition detection algorithms. It wasn't tested whether this has any effect on the previous versions.