Correctly check empty cuckoo move path
Correctly check that there are no pieces between the start and target square of the saved cuckoo move, using a new array LineBetween, which is just like Line but it excludes the start and target square.
Also changed the counter used from FiftyMovePlyCount to PlyCountReversible. The reason why is explained in a comment.
Finally, changed ply counter increment to 2 to save time, and to ensure only positions with the opponent's turn are considered.