Added option to not save children nodes
Now the reference to these nodes can optionally not be saved in the parent node, meaning once the search moves on, these are forgotten.
This significantly reduces the memory usage of the program to 400-500 MB during a very long search, compared to the >2 GB it was reaching before.
These changes also improved the engine's nodes per second during testing by upwards of 40%. More testing is needed to determine the increase in playing strength.