Fixed draw by repetition detection
Draws by repetition from positions before the root were previously not detected.
This is because the position history information used to detect draws by repetition was lost when cloning the board using ConvertFromFen on the current fen of the initial board.

The bug was likely introduced when the new system for keeping track of past positions, using GameStateHistory and PositionKeyHistory, was added.

Now, using the recently implemented Board.Clone method, all useful info about a position is preserved, including the PositionKeyHistory.